Handle two-column layouts
The `\circle` command takes a number, not a length, but it is being passed a length.  Ordinarily that doesn't seem to make a difference, but it breaks on short text in a two-column book-chapter.  Here's the MWE:

```
\documentclass[twocolumn]{book}
\usepackage{circledsteps}
\begin{document}
\chapter{Required}
Doesn't work: \Circled{i} % too short!

Works: \Circled{long}
\end{document}
```

Compile it and you'll see an error in the logs from pict2e.

The fix here is to convert the length `\csteps@YLength` to a number, using `\strip@pt`.